pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anatomize"
description = "Deterministic, token-efficient codebase packs and skeleton maps for AI review (Python)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dynamic = ["version"]
authors = [
    { name = "Bradley Max Segal", email = "bradleymaxsegal@gmail.com" }
]
keywords = ["ast", "code-analysis", "llm", "ai", "tree-sitter", "skeleton", "pack", "bundle"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "tree-sitter>=0.21,<0.22",
    "tree-sitter-python>=0.21,<0.22",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "tiktoken>=0.5",
    "tomli>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "types-PyYAML",
    "build",
    "twine",
]

[project.scripts]
anatomize = "anatomize.cli:app"

[project.urls]
Homepage = "https://github.com/BradSegal/code-skeleton"
Documentation = "https://github.com/BradSegal/code-skeleton#readme"
Repository = "https://github.com/BradSegal/code-skeleton"

[tool.hatch.build.targets.wheel]
packages = ["src/anatomize"]

[tool.hatch.version]
path = "src/anatomize/version.py"

[tool.black]
line-length = 119
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end CLI tests",
]
